[opentelemetry-exporter-azure-monitor] Changes for beta.1 of new package name (#12594)
Add api-extractor and cleanup public surface to match other languages. See Changelog for details.
