+++ sdk/monitor/opentelemetry-exporter-azure-monitor/package.json
@@ -13,7 +13,8 @@
     "build:browser": "echo skipped",
     "build:test": "echo skipped",
     "build:node": "tsc -p .",
-    "build": "npm run build:node && npm run build:browser",
+    "build": "npm run build:node && npm run build:browser && api-extractor run --local",
+    "extract-api": "tsc -p . && api-extractor run --local",
     "lint": "eslint . --ext .ts || exit 0",
     "test": "npm run test:node && npm run test:browser",
     "test:node": "npm run unit-test:node",
@@ -61,6 +62,7 @@
   },
   "devDependencies": {
     "@azure/eslint-plugin-azure-sdk": "^3.0.0",
+    "@microsoft/api-extractor": "7.7.11",
     "@types/mocha": "^7.0.2",
     "@types/node": "^10.0.0",
     "@typescript-eslint/eslint-plugin": "^2.0.0",
