+++ sdk/monitor/opentelemetry-exporter-azure-monitor/review/opentelemetry-exporter-azure-monitor.api.md
@@ -1,0 +1,29 @@
+## API Report File for "@microsoft/opentelemetry-exporter-azure-monitor"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+import { ExportResult } from '@opentelemetry/core';
+import { ReadableSpan } from '@opentelemetry/tracing';
+import { SpanExporter } from '@opentelemetry/tracing';
+
+// @public (undocumented)
+export interface AzureExporterConfig {
+    // (undocumented)
+    connectionString?: string;
+}
+
+// @public (undocumented)
+export class AzureMonitorTraceExporter implements SpanExporter {
+    constructor(options?: Partial<AzureExporterConfig>);
+    // (undocumented)
+    export(spans: ReadableSpan[], resultCallback: (result: ExportResult) => void): Promise<void>;
+    // (undocumented)
+    shutdown(): void;
+}
+
+
+// (No @packageDocumentation comment for this package)
+
+```
