+++ sdk/monitor/opentelemetry-exporter-azure-monitor/src/index.ts
@@ -1,4 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export * from "./export/trace";
+export { AzureMonitorTraceExporter } from "./export/trace";
+export { AzureExporterConfig } from "./config";
