+++ sdk/monitor/opentelemetry-exporter-azure-monitor/src/types.ts
@@ -1,8 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { ExportResult } from "@opentelemetry/core";
-import { ContextTagKeys, TelemetryItem as Envelope } from "./generated";
+import { ContextTagKeys } from "./generated";
 
 export type Tags = { [key in ContextTagKeys]: string };
 export type PropertyType = string | number | boolean | object | Array<PropertyType>;
@@ -12,14 +11,7 @@ export interface MSLink {
   id: string;
 }
 export type Measurements = { [key: string]: number };
-export type TelemetryProcessor = (envelope: Envelope) => boolean | void;
 export type SenderResult = { statusCode: number; result: string };
-
-export interface BaseExporter {
-  addTelemetryProcessor(processor: TelemetryProcessor): void;
-  clearTelemetryProcessors(): void;
-  exportEnvelopes(envelopes: Envelope[]): Promise<ExportResult>;
-}
 
 export interface Sender {
   send(payload: unknown[]): Promise<SenderResult>;
