+++ sdk/monitor/opentelemetry-exporter-azure-monitor/test/unit/export/export.test.ts
@@ -3,8 +3,7 @@
 
 import * as assert from "assert";
 import { ExportResult } from "@opentelemetry/core";
-import { AzureMonitorBaseExporter } from "../../../src/export/exporter";
-import { TelemetryProcessor } from "../../../src/types";
+import { AzureMonitorTraceExporter } from "../../../src/export/trace";
 import { DEFAULT_BREEZE_ENDPOINT } from "../../../src/Declarations/Constants";
 import {
   failedBreezeResponse,
@@ -20,15 +19,21 @@ function toObject<T>(obj: T): T {
 }
 
 describe("#AzureMonitorBaseExporter", () => {
-  class TestExporter extends AzureMonitorBaseExporter {
+  class TestExporter extends AzureMonitorTraceExporter {
+    private thisAsAny: any;
     constructor() {
       super({
-        instrumentationKey: "foo-ikey"
+        connectionString: `instrumentationkey=foo-ikey`
       });
+      this.thisAsAny = this;
     }
 
     getTelemetryProcesors() {
-      return this._telemetryProcessors;
+      return this.thisAsAny._telemetryProcessors;
+    }
+
+    async exportEnvelopesPrivate(payload: Envelope[]): Promise<ExportResult> {
+      return this.thisAsAny.exportEnvelopes(payload);
     }
   }
 
@@ -66,7 +71,7 @@ describe("#AzureMonitorBaseExporter", () => {
         const response = failedBreezeResponse(1, 429);
         scope.reply(429, JSON.stringify(response));
 
-        const result = await exporter.exportEnvelopes([envelope]);
+        const result = await exporter.exportEnvelopesPrivate([envelope]);
         assert.strictEqual(result, ExportResult.FAILED_RETRYABLE);
 
         const persistedEnvelopes = (await exporter["_persister"].shift()) as Envelope[];
@@ -79,7 +84,7 @@ describe("#AzureMonitorBaseExporter", () => {
         const response = partialBreezeResponse([200, 408, 408]);
         scope.reply(206, JSON.stringify(response));
 
-        const result = await exporter.exportEnvelopes([envelope, envelope, envelope]);
+        const result = await exporter.exportEnvelopesPrivate([envelope, envelope, envelope]);
         assert.strictEqual(result, ExportResult.FAILED_RETRYABLE);
 
         const persistedEnvelopes = (await exporter["_persister"].shift()) as Envelope[];
@@ -91,7 +96,7 @@ describe("#AzureMonitorBaseExporter", () => {
         const response = failedBreezeResponse(1, 400);
         scope.reply(400, JSON.stringify(response));
 
-        const result = await exporter.exportEnvelopes([envelope]);
+        const result = await exporter.exportEnvelopesPrivate([envelope]);
         assert.strictEqual(result, ExportResult.FAILED_NOT_RETRYABLE);
 
         const persistedEnvelopes = await exporter["_persister"].shift();
@@ -103,7 +108,7 @@ describe("#AzureMonitorBaseExporter", () => {
         const response = successfulBreezeResponse(1);
         scope.reply(200, JSON.stringify(response));
 
-        const result = await exporter.exportEnvelopes([envelope]);
+        const result = await exporter.exportEnvelopesPrivate([envelope]);
         assert.strictEqual(result, ExportResult.SUCCESS);
         assert.notStrictEqual(exporter["_retryTimer"], null);
 
@@ -117,100 +122,9 @@ describe("#AzureMonitorBaseExporter", () => {
         const response = successfulBreezeResponse(1);
         scope.reply(200, JSON.stringify(response));
 
-        const result = await exporter.exportEnvelopes([envelope]);
+        const result = await exporter.exportEnvelopesPrivate([envelope]);
         assert.strictEqual(result, ExportResult.SUCCESS);
         assert.strictEqual(exporter["_retryTimer"], "foo");
-      });
-    });
-  });
-
-  describe("Telemetry Processors", () => {
-    const nameProcessor: TelemetryProcessor = (envelope: Envelope) => {
-      envelope.name = "processor1";
-    };
-
-    const rejectProcessor: TelemetryProcessor = () => {
-      return false;
-    };
-
-    describe("#addTelemetryProcessor()", () => {
-      it("should add telemetry processors", () => {
-        const exporter = new TestExporter();
-        assert.strictEqual(exporter.getTelemetryProcesors().length, 0);
-
-        exporter.addTelemetryProcessor(nameProcessor);
-        assert.strictEqual(exporter.getTelemetryProcesors().length, 1);
-        assert.strictEqual(exporter.getTelemetryProcesors()[0], nameProcessor);
-
-        exporter.addTelemetryProcessor(rejectProcessor);
-        assert.strictEqual(exporter.getTelemetryProcesors().length, 2);
-        assert.strictEqual(exporter.getTelemetryProcesors()[0], nameProcessor);
-        assert.strictEqual(exporter.getTelemetryProcesors()[1], rejectProcessor);
-      });
-    });
-
-    describe("#clearTelemetryProcessors()", () => {
-      it("should clear all telemetry processors", () => {
-        const exporter = new TestExporter();
-        assert.strictEqual(exporter.getTelemetryProcesors().length, 0);
-
-        exporter.addTelemetryProcessor(nameProcessor);
-        assert.strictEqual(exporter.getTelemetryProcesors().length, 1);
-        assert.strictEqual(exporter.getTelemetryProcesors()[0], nameProcessor);
-
-        exporter.clearTelemetryProcessors();
-        assert.strictEqual(exporter.getTelemetryProcesors().length, 0);
-      });
-    });
-
-    describe("#_applyTelemetryProcessors()", () => {
-      it("should filter envelopes", () => {
-        const fooEnvelope = {
-          name: "foo",
-          time: new Date()
-        };
-        const barEnvelope = {
-          name: "bar",
-          time: new Date()
-        };
-
-        const exporter = new TestExporter();
-        assert.strictEqual(exporter.getTelemetryProcesors().length, 0);
-
-        exporter.addTelemetryProcessor((envelope) => {
-          return envelope.name === "bar";
-        });
-        const filtered = exporter["_applyTelemetryProcessors"]([fooEnvelope, barEnvelope]);
-        assert.strictEqual(filtered.length, 1);
-        assert.strictEqual(filtered[0], barEnvelope);
-      });
-
-      it("should filter modified envelopes", () => {
-        const fooEnvelope = {
-          name: "foo",
-          time: new Date()
-        };
-        const barEnvelope = {
-          name: "bar",
-          time: new Date()
-        };
-
-        const exporter = new TestExporter();
-        assert.strictEqual(exporter.getTelemetryProcesors().length, 0);
-
-        exporter.addTelemetryProcessor((envelope) => {
-          if (envelope.name === "bar") {
-            envelope.name = "baz";
-          }
-        });
-
-        exporter.addTelemetryProcessor((envelope) => {
-          return envelope.name === "baz";
-        });
-
-        const filtered = exporter["_applyTelemetryProcessors"]([fooEnvelope, barEnvelope]);
-        assert.strictEqual(filtered.length, 1);
-        assert.strictEqual(filtered[0].name, "baz");
       });
     });
   });
